requirements.txt
torch
pyyaml
argparse
tqdm
datasets
underthesea
transformers
